Add flutteroptions
